flask-toctou/pyproject.toml
[project]
name = "flask-toctou"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
]
